+++ core-ng/src/main/java/core/framework/json/JSON.java
@@ -2,7 +2,6 @@ package core.framework.json;
 
 import com.fasterxml.jackson.databind.JavaType;
 
-import javax.annotation.Nullable;
 import java.io.IOException;
 import java.io.UncheckedIOException;
 import java.lang.reflect.Type;
@@ -13,26 +12,29 @@ import static core.framework.internal.json.JSONMapper.OBJECT_MAPPER;
  * @author neo
  */
 public final class JSON {
-    @Nullable
     public static <T> T fromJSON(Type instanceType, String json) {
         try {
             JavaType javaType = OBJECT_MAPPER.getTypeFactory().constructType(instanceType);
-            return OBJECT_MAPPER.readValue(json, javaType);
+            T result = OBJECT_MAPPER.readValue(json, javaType);
+            if (result == null) throw new Error("invalid json value, value=" + json);   // not allow passing "null" as json value
+            return result;
         } catch (IOException e) {
             throw new UncheckedIOException(e);
         }
     }
 
-    @Nullable
     public static <T> T fromJSON(Class<T> instanceClass, String json) {
         try {
-            return OBJECT_MAPPER.readValue(json, instanceClass);
+            T result = OBJECT_MAPPER.readValue(json, instanceClass);
+            if (result == null) throw new Error("invalid json value, value=" + json);   // not allow passing "null" as json value
+            return result;
         } catch (IOException e) {
             throw new UncheckedIOException(e);
         }
     }
 
-    public static String toJSON(@Nullable Object instance) {
+    public static String toJSON(Object instance) {
+        if (instance == null) throw new Error("instance must not be null");
         try {
             return OBJECT_MAPPER.writeValueAsString(instance);
         } catch (IOException e) {